pyproject.toml
[tool.poetry]
name = "thema"
version = "0.1.0"
description = ""
authors = [
    "jeremy.wayland <jeremy.don.wayland@gmail.com>",
    "stuart.wayland <swayland@ucsc.edu>",
    "sid.gathrid <sgathrid@yahoo.com>",
]
readme = "README.md"
packages = [{ include = "thema" }]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.24.2"
networkx = "^3.0"
pandas = "^2.2.1"
jupyter = "^1.0.0"
omegaconf = "^2.3.0"
category-encoders = "^2.6.3"
tqdm = "^4.66.2"
termcolor = "^2.4.0"
umap-learn = "^0.5.5"
kmapper = "^2.0.1"
scikit-learn = "^1.4.1.post1"
hdbscan = "^0.8.33"
pot = "^0.9.3"
matplotlib = "^3.8.3"
gudhi = "^3.9.0"
plotly = "^5.20.0"
invoke = "^2.2.0"
grakel = "^0.1.10"
seaborn = "^0.13.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
openpyxl = "^3.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
